Broke out the find min and max dates
