+++ src/file-handling/create-output.ts
@@ -1,10 +1,11 @@
 import { AllRepoStatistics } from '../generator/models/all-repo-statistics';
 import { promisify } from 'util';
 import * as fs from 'fs';
-import moment from 'moment';
 import { DateCount } from '../generator/models/date-count';
 import { mappingToCumlative } from '../generator/mapping-to-cumulative';
 import { normaliseDates } from './normalise-dates';
+import { findMaxDate } from './find-max-date';
+import { findMinDate } from './find-min-date';
 
 const writeFile = promisify(fs.writeFile);
 
@@ -67,38 +68,4 @@ export async function createOutput(
 
     const csv = lines.map(line => line.join(',')).join('\n');
     await writeFile(filename, csv);
-}
-
-function findMinDate(allRepoStatistics: AllRepoStatistics): moment.Moment {
-    // @ts-ignore
-    let min: moment.Moment = null;
-    for (const repo in allRepoStatistics) {
-        if (allRepoStatistics.hasOwnProperty(repo)) {
-            const repoStatistic = allRepoStatistics[repo];
-            const keys = Object.keys(repoStatistic);
-            keys.map(key => moment(key)).forEach(date => {
-                if (min === null || date.isSameOrBefore(min)) {
-                    min = date;
-                }
-            });
-        }
-    }
-    return min;
-}
-
-function findMaxDate(allRepoStatistics: AllRepoStatistics): moment.Moment {
-    // @ts-ignore
-    let max: moment.Moment = null;
-    for (const repo in allRepoStatistics) {
-        if (allRepoStatistics.hasOwnProperty(repo)) {
-            const repoStatistic = allRepoStatistics[repo];
-            const keys = Object.keys(repoStatistic);
-            keys.map(key => moment(key)).forEach(date => {
-                if (max === null || date.isSameOrAfter(max)) {
-                    max = date;
-                }
-            });
-        }
-    }
-    return max;
 }
